+++ CHANGELOG.md
@@ -3,12 +3,13 @@ Changelog
 
 Notable changes will be documented in this file. The project adheres to [Semantic Versioning].
 
-Unreleased
-----------
+1.4.0
+-----
 
 * Improved `MailService`
 * Implemented `StringUtility`
 * Implemented custom `findByUid` finder in `AbstractRepository`
+* Minor changes
 
 1.3.0
 -----